fix crash when resizing window

+++ Scripts/Editor/Editor_Object.gd
@@ -42,50 +42,49 @@ func _process(delta):
 	#print(cursor_pos)
 	
 	object_point = WhatObject()
-	
 	#if (preview_level.get_child(0) != object_point.collider):
 	#	for n in preview_level.get_children():
 	#		preview_level.remove_child(n)
 	#		n.queue_free()
-	
-	if(play.edit_mode and can_place and object_point.collider.name == "floor"):
-		#var preview_item = preview.instance()
-		#preview_level.add_child(preview_item)
-		#preview_item.global_translation = cursor_pos
-		preview_item.set_visible(true)
-		delete_item.set_visible(false)
-		#print(cursor_pos)
-		preview_item.global_translation = cursor_pos
-	elif(play.edit_mode and can_place or "object" in object_point.collider.name):
-		preview_item.set_visible(false)
-		delete_item.set_visible(true)
-		cursor_pos.y = 2
-		delete_item.global_translation = cursor_pos
-	else:
-		preview_item.set_visible(false)
-		delete_item.set_visible(false)
-		
-	if (Input.is_action_just_pressed("mb_left")):
-		print(object_point.collider.name)
+	if(object_point.has("position")):
+		if(play.edit_mode and can_place and object_point.collider.name == "floor"):
+			#var preview_item = preview.instance()
+			#preview_level.add_child(preview_item)
+			#preview_item.global_translation = cursor_pos
+			preview_item.set_visible(true)
+			delete_item.set_visible(false)
+			#print(cursor_pos)
+			preview_item.global_translation = cursor_pos
+		elif(play.edit_mode and can_place or "object" in object_point.collider.name):
+			preview_item.set_visible(false)
+			delete_item.set_visible(true)
+			cursor_pos.y = 2
+			delete_item.global_translation = cursor_pos
+		else:
+			preview_item.set_visible(false)
+			delete_item.set_visible(false)
+			
+		if (Input.is_action_just_pressed("mb_left")):
+			print(object_point.collider.name)
+			pass
+			
+		if (Input.is_action_just_pressed("mb_right") and "object" in object_point.collider.name):
+			delete_object = level.get_node(object_point.collider.name)
+			print(delete_object)
+			delete_object.queue_free()
+			pass
+			
+		#Checks for Input
+		if(can_place and Input.is_action_just_pressed("mb_left") and current_item != null and object_point.collider.name == "floor"):
+			#	for n in preview_level.get_children():
+			#		preview_level.remove_child(n)
+			#		n.queue_free()
+			var new_item = current_item.instance() 
+			if (new_item != null):
+				level.add_child(new_item)
+				new_item.owner = level
+				new_item.global_translation = cursor_pos
 		pass
-		
-	if (Input.is_action_just_pressed("mb_right") and "object" in object_point.collider.name):
-		delete_object = level.get_node(object_point.collider.name)
-		print(delete_object)
-		delete_object.queue_free()
-		pass
-		
-	#Checks for Input
-	if(can_place and Input.is_action_just_pressed("mb_left") and current_item != null and object_point.collider.name == "floor"):
-		#	for n in preview_level.get_children():
-		#		preview_level.remove_child(n)
-		#		n.queue_free()
-		var new_item = current_item.instance() 
-		if (new_item != null):
-			level.add_child(new_item)
-			new_item.owner = level
-			new_item.global_translation = cursor_pos
-	pass
 	
 
 #Fires a ray to check for cursor 3D location
